Make size of unsigned int/log fully specific uint32_t and uint64_t

+++ papas/datatypes/DefinitionsCollections.h
@@ -19,7 +19,7 @@ class Edge;
 class Particle;
 
 typedef std::list<Particle> ListParticles;                   ///< list of Particles
-typedef std::unordered_map<unsigned long long, Edge> Edges;  ///< collection of Edge objects
+typedef std::unordered_map<uint64_t, Edge> Edges;  ///< collection of Edge objects
 #if WITHSORT
 typedef std::set<Identifier, std::greater<Identifier>> Ids;  ///< set containing Identifiers
 #else
